moving scripts to cli and changing utils to core
