+++ src/vtp/cli/accept_ballot.py
