+++ src/vtp/cli/create_blank_ballot.py
