+++ src/vtp/cli/generate_all_blank_ballots.py
