+++ src/vtp/cli/merge_contests.py
