+++ src/vtp/cli/setup_vtp_demo.py
