+++ src/vtp/cli/show_contest.py
