+++ src/vtp/cli/tally_contests.py
