+++ src/vtp/cli/verify_ballot_receipt.py
