+++ src/vtp/core/common.py
