+++ src/vtp/core/election_config.py
