+++ src/vtp/core/exceptions.py
