+++ src/vtp/ops/cast_ballot_operation.py
@@ -35,10 +35,10 @@ import traceback
 import pyinputplus
 
 # Local imports
-from vtp.utils.address import Address
-from vtp.utils.ballot import Ballot, BlankBallot, Contests
-from vtp.utils.common import Common, Globals, Shellout
-from vtp.utils.election_config import ElectionConfig
+from vtp.core.address import Address
+from vtp.core.ballot import Ballot, BlankBallot, Contests
+from vtp.core.common import Common, Globals, Shellout
+from vtp.core.election_config import ElectionConfig
 
 
 class CastBallotOperation:
