+++ src/vtp/ops/create_blank_ballot_operation.py
@@ -28,10 +28,10 @@ import logging
 import pprint
 
 # Local import
-from vtp.utils.address import Address
-from vtp.utils.ballot import BlankBallot
-from vtp.utils.common import Common
-from vtp.utils.election_config import ElectionConfig
+from vtp.core.address import Address
+from vtp.core.ballot import BlankBallot
+from vtp.core.common import Common
+from vtp.core.election_config import ElectionConfig
 
 
 class CreateBlankBallotOperation:
