+++ src/vtp/ops/generate_all_blank_ballots_operation.py
@@ -29,10 +29,10 @@ import os
 import pprint
 
 # Local import
-from vtp.utils.address import Address
-from vtp.utils.ballot import BlankBallot
-from vtp.utils.common import Common
-from vtp.utils.election_config import ElectionConfig
+from vtp.core.address import Address
+from vtp.core.ballot import BlankBallot
+from vtp.core.common import Common
+from vtp.core.election_config import ElectionConfig
 
 
 # pylint: disable=too-few-public-methods
