+++ src/vtp/ops/setup_vtp_demo_operation.py
@@ -29,8 +29,8 @@ import os
 import secrets
 
 # Local import
-from vtp.utils.common import Common, Globals, Shellout
-from vtp.utils.election_config import ElectionConfig
+from vtp.core.common import Common, Globals, Shellout
+from vtp.core.election_config import ElectionConfig
 
 
 class SetupVtpDemoOperation:
