+++ src/vtp/ops/show_contests_operation.py
@@ -29,8 +29,8 @@ import os
 import re
 
 # Local libraries
-from vtp.utils.common import Common, Globals, Shellout
-from vtp.utils.election_config import ElectionConfig
+from vtp.core.common import Common, Globals, Shellout
+from vtp.core.election_config import ElectionConfig
 
 
 class ShowContestsOperation:
