+++ src/vtp/ops/verify_ballot_receipt_operation.py
@@ -32,10 +32,10 @@ import os
 import re
 
 # Local import
-from vtp.utils.address import Address
-from vtp.utils.ballot import Ballot
-from vtp.utils.common import Common, Globals, Shellout
-from vtp.utils.election_config import ElectionConfig
+from vtp.core.address import Address
+from vtp.core.ballot import Ballot
+from vtp.core.common import Common, Globals, Shellout
+from vtp.core.election_config import ElectionConfig
 
 
 class VerifyBallotReceiptOperation:
